--- v1/backend/SPRING_LV_3_김영한_이동진_그리고/src/main/java/com/example/dongglee/MovieRepository.java
@@ -1,13 +1,0 @@
-package com.example.dongglee;
-
-import com.example.dongglee.domain.Movie;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface MovieRepository {
-	List<Movie> findAll();
-
-	Movie save(Movie movie);
-}
