+++ v1/backend/SPRING_LV_3_김영한_이동진_그리고/src/main/java/com/example/dongglee/controller/MovieController.java
@@ -1,29 +1,29 @@
-package com.example.dongglee;
+package com.example.dongglee.controller;
 
 import com.example.dongglee.domain.Movie;
 import com.example.dongglee.dto.MovieCreateRequestDto;
+import com.example.dongglee.service.MovieService;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
-
 @RestController
 @RequestMapping("/movies")
 @RequiredArgsConstructor
 public class MovieController {
 
-	private final MovieService movieService;
+    private final MovieService movieService;
 
-	@GetMapping
-	public List<Movie> getMovie() {
-		return movieService.getAllMovies();
-	}
+    @GetMapping
+    public List<Movie> getMovie() {
+        return movieService.getAllMovies();
+    }
 
-	@PostMapping
-	public Movie createMovie(MovieCreateRequestDto dto) {
-		return movieService.createMovie(dto);
-	}
+    @PostMapping
+    public Movie createMovie(MovieCreateRequestDto dto) {
+        return movieService.createMovie(dto);
+    }
 }
