+++ v1/backend/SPRING_LV_3_김영한_이동진_그리고/src/main/java/com/example/dongglee/repository/MovieRepository.java
@@ -1,0 +1,12 @@
+package com.example.dongglee.repository;
+
+import com.example.dongglee.domain.Movie;
+import java.util.List;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MovieRepository {
+    List<Movie> findAll();
+
+    Movie save(Movie movie);
+}
