+++ v1/backend/SPRING_LV_3_김영한_이동진_그리고/src/main/java/com/example/dongglee/service/MovieService.java
@@ -1,0 +1,29 @@
+package com.example.dongglee.service;
+
+import com.example.dongglee.domain.Movie;
+import com.example.dongglee.dto.MovieCreateRequestDto;
+import com.example.dongglee.repository.MovieRepository;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MovieService {
+
+    private final MovieRepository movieRepository;
+
+    public List<Movie> getAllMovies() {
+        return movieRepository.findAll();
+    }
+
+    public Movie createMovie(MovieCreateRequestDto dto) {
+        Movie movie = Movie.builder()
+                .director(dto.getDirector())
+                .title(dto.getTitle())
+                .filmedAt(dto.getFilmedAt())
+                .build();
+
+        return movieRepository.save(movie);
+    }
+}
